requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python-dotenv==0.10.1
six==1.15.0
SQLAlchemy==1.2.0
Werkzeug==1.0.1
